refactor: relocate KoColor domain models to core ritual package
This commit moves several ritual-related domain models from the feature-specific KoColor model package to a centralized core module. This refactoring ensures that shared data structures like `ClothingItem` and `SavedAnalysis` are accessible from a common location, improving modularity across the project.

- **`ComprehensiveLookBuilder.kt`**: Updated imports for `ClothingItem`, `CosmeticItem`, `HarmonizedLook`, and `MacroCategory` to point to `com.zoewave.probase.core.model.ritual`.
- **`WardrobeColorEngine.kt`**: Migrated `ClothingItem`, `SeasonalType`, and `Undertone` to the core ritual package.
- **`ColorScreen.kt`**: Updated UI-level references for `FashionAdvice`, `SavedAnalysis`, `SeasonalType`, and `Undertone`.
- **`ColorSearchViewModel.kt` & `ColorViewModel.kt`**: Refactored ViewModel dependencies to use the relocated core models.
